+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [3.3.1] - 2017-10-05
+
 ### Added
 - Support `kernel.Get<IEnumerable<IFoo>>()` [#252](https://github.com/ninject/Ninject/issues/252)